feat(rnmod): rename pkg
