+++ js/app/package.json
@@ -21,7 +21,7 @@
     "eslint": "^6.5.1",
     "jest": "^24.9.0",
     "metro-react-native-babel-preset": "^0.56.0",
-    "react-native-ipfs": "1.0.0",
+    "react-native-goipfs": "1.0.0",
     "react-test-renderer": "16.9.0",
     "wml": "^0.0.83"
   },
