+++ js/react-native-goipfs/.gitattributes
