+++ js/react-native-goipfs/.gitignore
