+++ js/react-native-goipfs/.watchmanconfig
