+++ js/react-native-goipfs/README.md
@@ -1,17 +1,17 @@
-# react-native-ipfs
+# react-native-goipfs
 
 ## Getting started
 
-`$ npm install react-native-ipfs --save`
+`$ npm install react-native-goipfs --save`
 
 ### Mostly automatic installation
 
-`$ react-native link react-native-ipfs`
+`$ react-native link react-native-goipfs`
 
 ## Usage
 
 ```javascript
-import IPFS from "react-native-ipfs";
+import IPFS from "react-native-goipfs";
 
 const ipfs = new IPFS();
