+++ js/react-native-goipfs/android/.gitignore
