+++ js/react-native-goipfs/android/build.gradle
