+++ js/react-native-goipfs/android/src/main/AndroidManifest.xml
