+++ js/react-native-goipfs/android/src/main/java/com/reactlibrary/IpfsModule.java
