+++ js/react-native-goipfs/android/src/main/java/com/reactlibrary/IpfsPackage.java
