+++ js/react-native-goipfs/index.js
