+++ js/react-native-goipfs/ios/Ipfs.h
