+++ js/react-native-goipfs/ios/Ipfs.m
