+++ js/react-native-goipfs/react-native-ipfs.podspec
@@ -3,18 +3,18 @@ require "json"
 package = JSON.parse(File.read(File.join(__dir__, "package.json")))
 
 Pod::Spec.new do |s|
-  s.name         = "react-native-ipfs"
+  s.name         = "react-native-goipfs"
   s.version      = package["version"]
   s.summary      = package["description"]
   s.description  = <<-DESC
-                  react-native-ipfs
+                  react-native-goipfs
                    DESC
-  s.homepage     = "https://github.com/github_account/react-native-ipfs"
+  s.homepage     = "https://github.com/github_account/react-native-goipfs"
   s.license      = "MIT"
   # s.license    = { :type => "MIT", :file => "FILE_LICENSE" }
   s.authors      = { "Your Name" => "yourname@email.com" }
   s.platforms    = { :ios => "9.0" }
-  s.source       = { :git => "https://github.com/github_account/react-native-ipfs.git", :tag => "#{s.version}" }
+  s.source       = { :git => "https://github.com/github_account/react-native-goipfs.git", :tag => "#{s.version}" }
 
   s.source_files = "ios/**/*.{h,m,swift}"
   s.requires_arc = true
